Fixes new CRUD

+++ db/migrate/20211123153241_create_wallets.rb
@@ -2,7 +2,7 @@ class CreateWallets < ActiveRecord::Migration[6.1]
   def change
     create_table :wallets do |t|
       t.string :descryption
-      t.int :wallet_budget
+      t.bigint :wallet_budget
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
